Add fip signal visualizer
